Update README.md
Dr. Adam,

I have been trying to do this on my own so I learn but I had a lot of issues with the merging. Stata keeps saying that it cannot find the variable, even though I saved the variable in both data sets under the same name, including spelling and capitalization. I am not sure what to do. I am also having issues opening the files for the merging from the internet; sometimes it works and sometimes the data set is all messed up.

Gabby